fix unchecked "useful links" item in drawer, rename package "dataclasses" to "models"
