+++ app/src/main/java/pro/dprof/dorprofzhelzszd/data/db/DbHelper.java
@@ -19,7 +19,7 @@ package pro.dprof.dorprofzhelzszd.data.db;
 
 import java.util.List;
 
-import pro.dprof.dorprofzhelzszd.dataclasses.Documents;
+import pro.dprof.dorprofzhelzszd.models.Documents;
 
 public interface DbHelper {
