+++ app/src/main/java/pro/dprof/dorprofzhelzszd/data/db/OpenDbHelper.java
@@ -25,7 +25,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import pro.dprof.dorprofzhelzszd.dataclasses.Documents;
+import pro.dprof.dorprofzhelzszd.models.Documents;
 import pro.dprof.dorprofzhelzszd.utils.Constants;
 
 final class OpenDbHelper extends SQLiteOpenHelper {
