+++ app/src/main/java/pro/dprof/dorprofzhelzszd/models/Documents.java
@@ -1,4 +1,4 @@
-package pro.dprof.dorprofzhelzszd.dataclasses;
+package pro.dprof.dorprofzhelzszd.models;
 
 import java.util.List;
