+++ app/src/main/java/pro/dprof/dorprofzhelzszd/models/Staff.java
@@ -1,4 +1,4 @@
-package pro.dprof.dorprofzhelzszd.dataclasses;
+package pro.dprof.dorprofzhelzszd.models;
 
 public final class Staff {
