+++ app/src/main/res/menu/menu_main_drawer.xml
@@ -23,7 +23,9 @@
 
     </group>
 
-    <group android:id="@+id/second_drawer_group">
+    <group
+        android:id="@+id/second_drawer_group"
+        android:checkableBehavior="single">
 
         <item
             android:id="@+id/item_drawer_useful_links"
